+++ supervisely/nn/benchmark/visualization/vis_click_data.py
@@ -1,0 +1,88 @@
+from __future__ import annotations
+
+from collections import defaultdict
+
+from supervisely.nn.benchmark.evaluation.object_detection.metric_provider import (
+    MetricProvider,
+)
+
+
+class IdMapper:
+    def __init__(self, coco_dataset: dict):
+        self.map_img = {x["id"]: x["sly_id"] for x in coco_dataset["images"]}
+        self.map_obj = {x["id"]: x["sly_id"] for x in coco_dataset["annotations"]}
+
+
+class ClickData:
+    def __init__(self, m: MetricProvider, gt_id_mapper: IdMapper, dt_id_mapper: IdMapper):
+        self.m = m
+        # self.m_full = m_full
+        self.gt_id_mapper = gt_id_mapper
+        self.dt_id_mapper = dt_id_mapper
+        self.catId2name = {cat_id: cat["name"] for cat_id, cat in m.cocoGt.cats.items()}
+
+        self.outcome_counts = {
+            "TP": self._gather_matches(self.m.tp_matches),
+            "FN": self._gather_matches(self.m.fn_matches),
+            "FP": self._gather_matches(self.m.fp_matches),
+        }
+
+        outcome_counts_by_class = defaultdict(lambda: {"TP": [], "FN": [], "FP": []})
+        for match in self.m.matches:
+            cat_id = match["category_id"]
+            cat_name = self.m.cocoGt.cats[cat_id]["name"]
+            outcome_counts_by_class[cat_name][match["type"]].append(self._gather(match))
+        self.outcome_counts_by_class = dict(outcome_counts_by_class)
+
+        self.objects_by_class = {cat_name: [] for cat_name in self.m.cat_names}  # ! ??? ШТОЭТА???
+        for match in self.m.matches:
+            cat_id = match["category_id"]
+            cat_name = self.m.cocoGt.cats[cat_id]["name"]
+            self.objects_by_class[cat_name].append(self._gather(match))
+
+        self.confusion_matrix = self._confusion_matrix()
+        self.frequently_confused = self._frequently_confused(self.confusion_matrix)
+
+    def _confusion_matrix(self):
+        confusion_matrix_ids = defaultdict(list)
+        none_name = "(None)"
+        for match in self.m.confused_matches:
+            cat_pred = self.catId2name[match["category_id"]]
+            cat_gt = self.catId2name[self.m.cocoGt.anns[match["gt_id"]]["category_id"]]
+            confusion_matrix_ids[(cat_pred, cat_gt)].append(self._gather(match))
+
+        for match in self.m.tp_matches:
+            cat_name = self.catId2name[match["category_id"]]
+            confusion_matrix_ids[(cat_name, cat_name)].append(self._gather(match))
+
+        for match in self.m.fp_not_confused_matches:
+            cat_pred = self.catId2name[match["category_id"]]
+            confusion_matrix_ids[(cat_pred, none_name)].append(self._gather(match))
+
+        for match in self.m.fn_matches:
+            cat_gt = self.catId2name[match["category_id"]]
+            confusion_matrix_ids[(none_name, cat_gt)].append(self._gather(match))
+        return confusion_matrix_ids
+
+    def _frequently_confused(self, confusion_matrix_ids: dict):
+        cm = self.m.confusion_matrix()
+        fcp = self.m.frequently_confused(cm)
+        pairs = fcp["category_pair"]
+        frequently_confused = {}
+        for i, pair in enumerate(pairs):
+            cat_a, cat_b = pair
+            joint = confusion_matrix_ids[(cat_a, cat_b)] + confusion_matrix_ids[(cat_b, cat_a)]
+            joint = sorted(joint, key=lambda x: x["gt_img_id"])
+            frequently_confused[pair] = joint
+        return frequently_confused
+
+    def _gather(self, match: dict):
+        return {
+            "gt_img_id": self.gt_id_mapper.map_img[match["image_id"]],
+            "dt_img_id": self.dt_id_mapper.map_img[match["image_id"]],
+            "gt_obj_id": self.gt_id_mapper.map_obj.get(match["gt_id"]),
+            "dt_obj_id": self.dt_id_mapper.map_obj.get(match["dt_id"]),
+        }
+
+    def _gather_matches(self, matches: list):
+        return [self._gather(d) for d in matches]
