+++ supervisely/nn/benchmark/visualization/visualizer.py
@@ -1,25 +1,21 @@
 from __future__ import annotations
 
 import json
-import os
 import pickle
-from collections import defaultdict
-from typing import TYPE_CHECKING, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Tuple
 
 import pandas as pd
 import ujson
 from jinja2 import Template
-from pycocotools.coco import COCO
 
-from supervisely._utils import *
 from supervisely.annotation.annotation import Annotation
 from supervisely.annotation.tag import Tag
 from supervisely.annotation.tag_meta import TagApplicableTo, TagMeta, TagValueType
-from supervisely.api.api import Api
 from supervisely.app.widgets import GridGalleryV2
 from supervisely.convert.image.coco.coco_helper import HiddenCocoPrints
 from supervisely.geometry.rectangle import Rectangle
-from supervisely.io.fs import *
+from supervisely.io.fs import file_exists, mkdir
+from supervisely.nn.benchmark.cv_tasks import CVTask
 from supervisely.project.project import Dataset, OpenMode, Project
 from supervisely.project.project_meta import ProjectMeta
 
@@ -29,198 +25,13 @@ if TYPE_CHECKING:
 from supervisely.nn.benchmark.evaluation.object_detection.metric_provider import (
     MetricProvider,
 )
-from supervisely.nn.benchmark.visualization.visualizations import *
-from supervisely.nn.benchmark.visualization.visualizations import MetricVis
+from supervisely.nn.benchmark.visualization.vis_click_data import ClickData, IdMapper
+from supervisely.nn.benchmark.visualization.vis_metrics import ALL_METRICS, MetricVis
+from supervisely.nn.benchmark.visualization.vis_templates import generate_main_template
+from supervisely.nn.benchmark.visualization.vis_widgets import Widget
 from supervisely.project.project_meta import ProjectMeta
 from supervisely.sly_logger import logger
 from supervisely.task.progress import tqdm_sly
-
-_METRIC_VISUALIZATIONS = (
-    Overview,
-    ExplorerGrid,
-    ModelPredictions,
-    # # WhatIs,
-    OutcomeCounts,
-    Recall,
-    Precision,
-    RecallVsPrecision,
-    PRCurve,
-    PRCurveByClass,
-    ConfusionMatrix,
-    FrequentlyConfused,
-    IOUDistribution,
-    ReliabilityDiagram,
-    ConfidenceScore,
-    F1ScoreAtDifferentIOU,
-    ConfidenceDistribution,
-    PerClassAvgPrecision,
-    PerClassOutcomeCounts,
-    # segmentation-only
-    # # TODO integrate binary files while saving to self.layout_dir to the current solution
-    # OverallErrorAnalysis,
-    # ClasswiseErrorAnalysis,
-)
-
-
-def generate_main_template(metric_visualizations: List[MetricVis]):
-    template_str = """<div>
-    <sly-iw-sidebar :options="{ height: 'calc(100vh - 130px)', clearMainPanelPaddings: true, leftSided: false,  disableResize: true, sidebarWidth: 300 }">
-        <div slot="sidebar">"""
-
-    for vis in metric_visualizations:
-        template_str += vis.template_sidebar_str
-
-    template_str += """\n        </div>
-      
-        <div style="padding: 0 15px;">"""
-
-    for vis in metric_visualizations:
-        template_str += vis.template_main_str
-
-    template_str += "\n        </div>\n    </sly-iw-sidebar>"
-
-    template_str += """\n
-        <sly-iw-gallery
-            ref='modal_general'
-            iw-widget-id='modal_general'
-            :options="{'isModalWindow': true}"
-            :actions="{
-                'init': {
-                'dataSource': '/data/modal_general.json',
-                },
-            }"
-            :command="command"
-            :data="data"
-        /> \n
-        <sly-iw-gallery
-            ref='modal_extra'
-            iw-widget-id='modal_extra'
-            :options="{'isModalWindow': true}"
-            :actions="{
-                'init': {
-                'dataSource': '/data/modal_extra.json',
-                },
-            }"
-            :command="command"
-            :data="data"
-        /> \n
-        <sly-iw-gallery
-            ref='modal_general_diff'
-            iw-widget-id='modal_general_diff'
-            :options="{'isModalWindow': true}"
-            :actions="{
-                'init': {
-                'dataSource': '/data/modal_general_diff.json',
-                },
-            }"
-            :command="command"
-            :data="data"
-        /> \n
-        <sly-iw-gallery
-            ref='modal_general_diff'
-            iw-widget-id='modal_general_diff'
-            :options="{'isModalWindow': true}"
-            :actions="{
-                'init': {
-                'dataSource': '/data/modal_general_diff.json',
-                },
-                'chart-click': {
-                    'dataSource': '/data/gallery_explorer_grid_diff_data.json',
-                    'galleryId': 'modal_extra',
-                        'limit': 3,
-                },
-            }"
-            :command="command"
-            :data="data"
-        > \n
-            <span slot="image-left-header">
-                <i class="zmdi zmdi-collection-image"></i> More details
-            </span>
-        </sly-iw-gallery>
-        </div>"""
-
-    return template_str
-
-
-class IdMapper:
-    def __init__(self, coco_dataset: dict):
-        self.map_img = {x["id"]: x["sly_id"] for x in coco_dataset["images"]}
-        self.map_obj = {x["id"]: x["sly_id"] for x in coco_dataset["annotations"]}
-
-
-class ClickData:
-    def __init__(self, m: MetricProvider, gt_id_mapper: IdMapper, dt_id_mapper: IdMapper):
-        self.m = m
-        # self.m_full = m_full
-        self.gt_id_mapper = gt_id_mapper
-        self.dt_id_mapper = dt_id_mapper
-        self.catId2name = {cat_id: cat["name"] for cat_id, cat in m.cocoGt.cats.items()}
-
-        self.outcome_counts = {
-            "TP": self._gather_matches(self.m.tp_matches),
-            "FN": self._gather_matches(self.m.fn_matches),
-            "FP": self._gather_matches(self.m.fp_matches),
-        }
-
-        outcome_counts_by_class = defaultdict(lambda: {"TP": [], "FN": [], "FP": []})
-        for match in self.m.matches:
-            cat_id = match["category_id"]
-            cat_name = self.m.cocoGt.cats[cat_id]["name"]
-            outcome_counts_by_class[cat_name][match["type"]].append(self._gather(match))
-        self.outcome_counts_by_class = dict(outcome_counts_by_class)
-
-        self.objects_by_class = {cat_name: [] for cat_name in self.m.cat_names}  # ! ??? ШТОЭТА???
-        for match in self.m.matches:
-            cat_id = match["category_id"]
-            cat_name = self.m.cocoGt.cats[cat_id]["name"]
-            self.objects_by_class[cat_name].append(self._gather(match))
-
-        self.confusion_matrix = self._confusion_matrix()
-        self.frequently_confused = self._frequently_confused(self.confusion_matrix)
-
-    def _confusion_matrix(self):
-        confusion_matrix_ids = defaultdict(list)
-        none_name = "(None)"
-        for match in self.m.confused_matches:
-            cat_pred = self.catId2name[match["category_id"]]
-            cat_gt = self.catId2name[self.m.cocoGt.anns[match["gt_id"]]["category_id"]]
-            confusion_matrix_ids[(cat_pred, cat_gt)].append(self._gather(match))
-
-        for match in self.m.tp_matches:
-            cat_name = self.catId2name[match["category_id"]]
-            confusion_matrix_ids[(cat_name, cat_name)].append(self._gather(match))
-
-        for match in self.m.fp_not_confused_matches:
-            cat_pred = self.catId2name[match["category_id"]]
-            confusion_matrix_ids[(cat_pred, none_name)].append(self._gather(match))
-
-        for match in self.m.fn_matches:
-            cat_gt = self.catId2name[match["category_id"]]
-            confusion_matrix_ids[(none_name, cat_gt)].append(self._gather(match))
-        return confusion_matrix_ids
-
-    def _frequently_confused(self, confusion_matrix_ids: dict):
-        cm = self.m.confusion_matrix()
-        fcp = self.m.frequently_confused(cm)
-        pairs = fcp["category_pair"]
-        frequently_confused = {}
-        for i, pair in enumerate(pairs):
-            cat_a, cat_b = pair
-            joint = confusion_matrix_ids[(cat_a, cat_b)] + confusion_matrix_ids[(cat_b, cat_a)]
-            joint = sorted(joint, key=lambda x: x["gt_img_id"])
-            frequently_confused[pair] = joint
-        return frequently_confused
-
-    def _gather(self, match: dict):
-        return {
-            "gt_img_id": self.gt_id_mapper.map_img[match["image_id"]],
-            "dt_img_id": self.dt_id_mapper.map_img[match["image_id"]],
-            "gt_obj_id": self.gt_id_mapper.map_obj.get(match["gt_id"]),
-            "dt_obj_id": self.dt_id_mapper.map_obj.get(match["dt_id"]),
-        }
-
-    def _gather_matches(self, matches: list):
-        return [self._gather(d) for d in matches]
 
 
 class Visualizer:
@@ -248,9 +59,10 @@ class Visualizer:
         if benchmark.cv_task == CVTask.OBJECT_DETECTION:
             self._initialize_object_detection_loader()
         else:
-            raise NotImplementedError("Please specify a new CVTask")
+            raise NotImplementedError(f"CV task {benchmark.cv_task} is not supported yet")
 
     def _initialize_object_detection_loader(self):
+        from pycocotools.coco import COCO
 
         cocoGt_path, cocoDt_path, eval_data_path, inference_info_path = (
             self.eval_dir + "/cocoGt.json",
@@ -282,15 +94,14 @@ class Visualizer:
         else:
             self.inference_info = self._benchmark._eval_inference_info
 
-        mp = MetricProvider(
+        self.mp = MetricProvider(
             eval_data["matches"],
             eval_data["coco_metrics"],
             eval_data["params"],
             cocoGt,
             cocoDt,
         )
-        mp.calculate()
-        self.mp = mp
+        self.mp.calculate()
 
         self.df_score_profile = pd.DataFrame(
             self.mp.confidence_score_profile(), columns=["scores", "precision", "recall", "f1"]
@@ -309,30 +120,16 @@ class Visualizer:
         self.click_data = ClickData(self.mp.m, gt_id_mapper, dt_id_mapper)
         self.base_metrics = self.mp.base_metrics
 
-        datasets = self._api.dataset.get_list(self.dt_project_info.id)
-        tmp = {}
-        self.dt_images_dct = {}
-        self.dt_images_dct_by_name = {}
-        for d in datasets:
-            images = self._api.image.get_list(d.id)
-            tmp[d.id] = [x.id for x in images]
-            for info in images:
-                self.dt_images_dct[info.id] = info
-                self.dt_images_dct_by_name[info.name] = info
-
-        self.dt_ann_jsons = {
-            ann.image_id: ann.annotation
-            for d in datasets
-            for ann in self._api.annotation.download_batch(d.id, tmp[d.id])
-        }
-
         self._update_pred_dcts()
+        self._update_gt_dcts()
         self._update_diff_dcts()
+
+        self._objects_bindings = []
 
     def visualize(self):
         mkdir(f"{self.layout_dir}/data", remove_content_if_exists=True)
 
-        initialized = [mv(self) for mv in _METRIC_VISUALIZATIONS]
+        initialized = [mv(self) for mv in ALL_METRICS]
         initialized = [mv for mv in initialized if self.cv_task.value in mv.cv_tasks]
         for mv in initialized:
             for widget in mv.schema:
@@ -349,10 +146,6 @@ class Visualizer:
         gallery._update_filters()
         res.update(gallery.get_json_state())
 
-        # res.update(gallery.get_json_data()["content"])
-        # res["layoutData"] = res.pop("annotations")
-
-        # res["projectMeta"] = self.dt_project_meta.to_json()
         self.dt_project_meta = ProjectMeta.from_json(
             data=self._api.project.get_meta(id=self.dt_project_info.id)
         )
@@ -496,90 +289,60 @@ class Visualizer:
         logger.info("Saved: %r", "state.json")
 
     def update_diff_annotations(self):
-        self._add_tags_to_dt_project(self.mp.matches, self.dt_project_info.id)
-        gt_project_path, dt_project_path = self._benchmark._download_projects()
+        meta = self._update_pred_meta_with_tags(self.dt_project_info.id, self.dt_project_meta)
+        self._api.project.update_meta(self.diff_project_info.id, meta)
+
+        self.dt_project_meta = meta
+        self._add_tags_to_pred_project(self.mp.matches, self.dt_project_info.id)
+        gt_project_path, pred_project_path = self._benchmark._download_projects()
 
         gt_project = Project(gt_project_path, OpenMode.READ)
-        dt_project = Project(dt_project_path, OpenMode.READ)
-        dt_project_meta = self.dt_project_meta
+        pred_project = Project(pred_project_path, OpenMode.READ)
 
-        dt_images_dct = {}
-        dt_anns_dct = {}
-        names_dct = {}
+        names_map = {}
+        pred_images_map = {}
+        pred_anns_map = {}
 
-        for dataset in dt_project.datasets:
+        for dataset in pred_project.datasets:
             dataset: Dataset
-            paths = list_files(dt_project_path + f"/{dataset.name}/img_info")
-            infos = [ImageInfo(**json.load(open(path, "r"))) for path in paths]
-            image_names = [x.name for x in sorted(infos, key=lambda info: info.id)]
+            image_names = dataset.get_items_names()
+            infos = [dataset.get_item_info(name) for name in image_names]
 
-            dt_anns_dct[dataset.name] = [
-                dataset.get_ann(name, dt_project_meta) for name in image_names
-            ]
-            dt_images_dct[dataset.name] = [dataset.get_image_info(name) for name in image_names]
-            names_dct[dataset.name] = image_names
+            pred_images_map[dataset.name] = infos
+            names_map[dataset.name] = image_names
+            pred_anns_map[dataset.name] = [dataset.get_ann(name, meta) for name in image_names]
 
-        gt_anns_dct = {}
-        for dataset in gt_project.datasets:
-            gt_anns_dct[dataset.name] = [
-                dataset.get_ann(name, gt_project.meta) for name in names_dct[dataset.name]
-            ]
-
-        # matched_ids = []
-        # for dataset in dt_project.datasets:
-        #     for dt_ann in dt_anns_dct[dataset.name]:
-        #         for label in dt_ann.labels:
-        #             matched_gt_id = label.tags.get("matched_gt_id")
-        #             if matched_gt_id is not None:
-        #                 matched_ids.append(matched_gt_id.value)
+        gt_anns_map = {}
+        for d in gt_project.datasets:
+            gt_anns_map[d.name] = [d.get_ann(name, gt_project.meta) for name in names_map[d.name]]
 
         matched_id_map = self._get_matched_id_map()  # dt_id -> gt_id
         matched_gt_ids = set(matched_id_map.values())
 
-        new_tag = TagMeta(
-            "outcome",
-            value_type=TagValueType.ONEOF_STRING,
-            possible_values=["TP", "FP", "FN"],
-            applicable_to=TagApplicableTo.OBJECTS_ONLY,
-        )
-        tag_metas = dt_project_meta.tag_metas
-        if dt_project_meta.get_tag_meta(new_tag.name) is None:
-            tag_metas = dt_project_meta.tag_metas.add(new_tag)
-
-        tag_metas = dt_project.meta.tag_metas
-        if dt_project.meta.get_tag_meta(new_tag.name) is None:
-            tag_metas = dt_project.meta.tag_metas.add(new_tag)
-
-        diff_meta = ProjectMeta(
-            obj_classes=gt_project.meta.obj_classes,
-            tag_metas=tag_metas,
-        )
-
-        self._api.project.update_meta(self.diff_project_info.id, diff_meta.to_json())
-        self._api.project.update_meta(self.dt_project_info.id, diff_meta.to_json())
+        outcome_tag = meta.get_tag_meta("outcome")
+        conf_meta = meta.get_tag_meta("confidence")
 
         with tqdm_sly(
-            desc="Creating diff_project", total=sum([len(x) for x in gt_anns_dct.values()])
+            desc="Creating diff_project", total=sum([len(x) for x in gt_anns_map.values()])
         ) as pbar1:
             with tqdm_sly(
-                desc="Updating dt_project", total=sum([len(x) for x in gt_anns_dct.values()])
+                desc="Updating pred_project", total=sum([len(x) for x in gt_anns_map.values()])
             ) as pbar2:
 
                 for dataset in self._api.dataset.get_list(self.diff_project_info.id):
                     diff_anns_new, dt_anns_new = [], []
 
-                    for gt_ann, dt_ann in zip(gt_anns_dct[dataset.name], dt_anns_dct[dataset.name]):
+                    for gt_ann, dt_ann in zip(
+                        gt_anns_map[dataset.name], pred_anns_map[dataset.name]
+                    ):
                         labels = []
                         for label in dt_ann.labels:
                             # match_tag_id = label.tags.get("matched_gt_id")
                             match_tag_id = matched_id_map.get(label.geometry.sly_id)
 
-                            if match_tag_id is not None:
-                                new = label.clone(tags=label.tags.add(Tag(new_tag, "TP")))
-                            else:
-                                new = label.clone(tags=label.tags.add(Tag(new_tag, "FP")))
-
-                            labels.append(new)
+                            value = "TP" if match_tag_id else "FP"
+                            label = label.add_tag(Tag(outcome_tag, value))
+                            labels.append(label)
 
                         dt_anns_new.append(Annotation(gt_ann.img_size, labels))
 
@@ -587,78 +350,94 @@ class Visualizer:
                             if label.geometry.sly_id not in matched_gt_ids and isinstance(
                                 label.geometry, Rectangle
                             ):
-                                conf_meta = dt_project_meta.get_tag_meta("confidence")
-
-                                labels.append(
-                                    label.clone(
-                                        tags=label.tags.add_items(
-                                            [Tag(new_tag, "FN"), Tag(conf_meta, 1)]
-                                        )
-                                    )
-                                )
+                                label = label.add_tags([Tag(outcome_tag, "FN"), Tag(conf_meta, 1)])
+                                labels.append(label)
 
                         diff_anns_new.append(Annotation(gt_ann.img_size, labels))
 
-                dt_image_ids = [x.id for x in dt_images_dct[dataset.name]]
-                self._api.annotation.upload_anns(dt_image_ids, dt_anns_new, progress_cb=pbar2)
+                    pred_img_ids = [x.id for x in pred_images_map[dataset.name]]
+                    self._api.annotation.upload_anns(pred_img_ids, dt_anns_new, progress_cb=pbar2)
 
-                diff_images = self._api.image.copy_batch(dataset.id, dt_image_ids)
+                    diff_images = self._api.image.copy_batch(dataset.id, pred_img_ids)
 
-                diff_image_ids = [image.id for image in diff_images]
-                self._api.annotation.upload_anns(diff_image_ids, diff_anns_new, progress_cb=pbar1)
+                    diff_img_ids = [image.id for image in diff_images]
+                    self._api.annotation.upload_anns(diff_img_ids, diff_anns_new, progress_cb=pbar1)
 
-        self._update_pred_dcts()
+        self._update_gt_dcts()
         self._update_diff_dcts()
+        self._update_pred_dcts()
 
-    def _update_pred_dcts(self):
+    def _update_gt_dcts(self):
         datasets = self._api.dataset.get_list(self.gt_project_info.id)
-        tmp = {}
         self.gt_images_dct = {}
         self.gt_images_dct_by_name = {}
         for d in datasets:
             images = self._api.image.get_list(d.id)
-            tmp[d.id] = [x.id for x in images]
             for info in images:
                 self.gt_images_dct[info.id] = info
                 self.gt_images_dct_by_name[info.name] = info
 
     def _update_diff_dcts(self):
         datasets = self._api.dataset.get_list(self.diff_project_info.id)
-        tmp = {}
         self.diff_images_dct = {}
         self.diff_images_dct_by_name = {}
         for d in datasets:
             images = self._api.image.get_list(d.id)
-            tmp[d.id] = [x.id for x in images]
             for info in images:
                 self.diff_images_dct[info.id] = info
                 self.diff_images_dct_by_name[info.name] = info
 
-    def _add_tags_to_dt_project(self, matches: list, dt_project_id: int):
-        api = self._api
-        match_tag_meta = TagMeta(
-            "matched_gt_id", TagValueType.ANY_NUMBER, applicable_to=TagApplicableTo.OBJECTS_ONLY
+    def _update_pred_dcts(self):
+        datasets = self._api.dataset.get_list(self.dt_project_info.id)
+        self.dt_images_dct = {}
+        self.dt_images_dct_by_name = {}
+        self.dt_ann_jsons = {}
+        for d in datasets:
+            images = self._api.image.get_list(d.id)
+            for info in images:
+                self.dt_images_dct[info.id] = info
+                self.dt_images_dct_by_name[info.name] = info
+
+            dt_anns = self._api.annotation.download_batch(d.id, [x.id for x in images])
+            self.dt_ann_jsons.update({ann.image_id: ann.annotation for ann in dt_anns})
+
+    def _update_pred_meta_with_tags(self, project_id: int, meta: ProjectMeta) -> ProjectMeta:
+        old_meta = meta
+        outcome_tag = TagMeta(
+            "outcome",
+            value_type=TagValueType.ONEOF_STRING,
+            possible_values=["TP", "FP", "FN"],
+            applicable_to=TagApplicableTo.OBJECTS_ONLY,
         )
-        iou_tag_meta = TagMeta(
-            "iou", TagValueType.ANY_NUMBER, applicable_to=TagApplicableTo.OBJECTS_ONLY
+        match_tag = TagMeta(
+            "matched_gt_id",
+            TagValueType.ANY_NUMBER,
+            applicable_to=TagApplicableTo.OBJECTS_ONLY,
+        )
+        iou_tag = TagMeta(
+            "iou",
+            TagValueType.ANY_NUMBER,
+            applicable_to=TagApplicableTo.OBJECTS_ONLY,
         )
 
-        # update project meta with new tag metas
-        meta = api.project.get_meta(dt_project_id)
-        meta = ProjectMeta.from_json(meta)
-        meta_old = meta
-        if not meta.tag_metas.has_key("matched_gt_id"):
-            meta = meta.add_tag_meta(match_tag_meta)
-        if not meta.tag_metas.has_key("iou"):
-            meta = meta.add_tag_meta(iou_tag_meta)
-        if meta != meta_old:
-            meta = api.project.update_meta(dt_project_id, meta)
-            self.dt_project_meta = meta
+        for tag in [outcome_tag, match_tag, iou_tag]:
+            if meta.get_tag_meta(tag.name) is None:
+                meta = meta.add_tag_meta(tag)
+
+        if old_meta == meta:
+            return meta
+
+        meta = self._api.project.update_meta(project_id, meta)
+        return meta
+
+        # conf_meta = meta.get_tag_meta("confidence")
+
+    def _add_tags_to_pred_project(self, matches: list, pred_project_id: int):
 
         # get tag metas
-        # outcome_tag_meta = meta.get_tag_meta("outcome")
-        match_tag_meta = meta.get_tag_meta("matched_gt_id")
-        iou_tag_meta = meta.get_tag_meta("iou")
+        # outcome_tag_meta = self.dt_project_meta.get_tag_meta("outcome")
+        match_tag_meta = self.dt_project_meta.get_tag_meta("matched_gt_id")
+        iou_tag_meta = self.dt_project_meta.get_tag_meta("iou")
 
         # mappings
         gt_ann_mapping = self.click_data.gt_id_mapper.map_obj
@@ -702,7 +481,7 @@ class Visualizer:
                     raise ValueError(f"Unknown match type: {match['type']}")
 
                 pbar.update(1)
-        api.image.tag.add_to_objects(dt_project_id, pred_tag_list)
+        self._api.image.tag.add_to_objects(pred_project_id, pred_tag_list)
 
     def _get_matched_id_map(self):
         gt_ann_mapping = self.click_data.gt_id_mapper.map_obj
